Highlight the current LSP signature help parameter
If you are using the language server protocol plugin in Neovim, and you
are inside of a function signature, triggering the signature help will
now highlight the function parameter you are currently editing.

Think 'foobar(bax, s|)' if your cursor is at | then you will now get
highlighting for the type of the second parameter of the function
"foobar". The highlighting uses the same colors as search.