formatting: improve grouping of colors and add comments
This does not affect the actual color scheme.

Make it easier to see which part of the color scheme is about builtin
functionality, and which is about 3rd party plugins.

I also made the plugin header comments a bit more uniform, because
apparently I'm looking for ways to procrastinate on the important
things in life.