Build the theme with a Lua template
Add a light weight Lua templating system to generate `colors/yui.vim`
and `doc/yui.txt`. The templating system uses hluv to adjust lightness
of a small set of palette colors, instead of hard coding everything left
and right. This will make it a lot easier to give the theme a more
coherent and consistent feel.

The templating system also generates the docs. Previously, the docs were
inconsistent in their use of help text directives and formatting. It'll
also ease maintenance if more things are generated automatically (less
things that need to be kept in sync manually).

Other changes:
* Add a Nix Flake which also creates a Vim plugin with the Nixpkgs
  infrastructure
* Improve the theme itself
  * Fix bugs in Lightline HL groups where the ctermbg and ctermfg codes
    were swapped
  * Add a few missing HL groups
  * Use more colors from the Dieter Rams color palette, which was the
    original idea anyway
  * Add fugitive HL groups
  * Add help text HL groups
* Add license and contributing information
* Add a Makefile
* Overhaul the README