Improve Lightline colors
The current lightline colors are way too dark. This commit generally
lightens the colors, so that we can then darken them as we go from
center to the sides.

First, we find the StatusLine guibg and lighten it a bit.

Then we define 5 shades, from middle to sides. The shades are based on
the base colors, which are StatusLine (lightened) and c.black1.

Co-authored-by: mambusskruj <mambus.skruj@gmail.com>