Add Lua conditionals for Vim script
In the past, it was awkward to make certian highlight blocks
conditional. For example, certain HL groups might only exist in Neovim
but not in Vim.

I initially added a "features" field that was a bit of hack. The field
was wrapped in a Vim condition, so it was a rather opaque string and the
concept wasn't reusable.

In an attempt to double down on the Lua-to-Vim approach, I created
combinators for And and Or. Using data instead of functions means you
can operate on these combinators with normal Lua code. For example, you
can run a function over all conditions in a conditional. This is done to
convert a conditional referencing Neovim/Vim feature strings into a
conditional that wraps these feature strings in 'has($feature)'.

These combinators can be reused in different contexts. At least I hope
so.