Various improvements
I've accumulated a lot of improvements in many different areas. In the
future I should really commit them as soon as I feel confident about
them but the problem is that sometimes you need to see several changes
in conjunction to know if they work.

These changes mostly fall in various buckets:
- Fixes for things that looked broken like the PmenuExtra not having a
  background color
- Colors that were fixed but which should have been computed relatively
- Give code more structure so it's easier to glance; this involved A LOT
  OF experimenting with bold, underline, and so on