Make Special bold
This probably needs some selective opt out for specific groups later on,
since a lot of groups are likely linked to Special that I'm not aware
of. The main motivating factor is that Janet (and probably other Lisps)
use Special for symbols and those are heavily used for certain DSLs:

  (loop [x :in bar :let [a 0]})

In this snippet the ':in' and ':let' use Special and I'd like those to
stand out.