General improvements
Until now, both dark and light themes had the same surfaces. They had
the base canvas, which provided e.g., the Neovim `Normal` foreground and
background colors. And then they had layers 1-3 which were increasingly
bright/light colored in the dark theme, and darker for the light theme.

I used the layers for things like `WinSeparator` and the status line.
But that ended up not looking great. Especially for a dark theme, having
too much bright stuff on the screen is very disorienting.

I want the status line and the window separator to be just slightly
darker in *both light and dark*. They have to be decoupled from the
normal layers.

So this commit adds some custom code to `nvim/theme.lua` which uses some
`dsl` functions to generate a status line that is, relative to `Normal`
of that (dark) theme variant, darker.

Lastly, the Treesitter context window is less intrusive now.

And floating windows now use the @markup.raw.block.markdown background,
since most of them actually render Markdown code blocks. I'll have to
see how that works out. In practice, it's really hard to get
consistently good looking floating windows. I think you'd need actual
cascading overrides for this to force the `guibg` of everything
displayed in a floating window to be of some uniform value.